backend/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "wms-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "greenlet>=3.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "bcrypt>=4.1",
  "pyjwt>=2.8",
  "email-validator>=2.1",
  "celery[redis]>=5.3",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4",
  "mypy>=1.10",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "types-requests>=2.32.0.20240622",
  "aiosqlite>=0.20",
]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = ["alembic/versions", "alembic/env.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
# DATABASE_URL must be set before importing app.db.session (shared engine for HTTP + BackgroundTasks).
"tests/conftest.py" = ["E402"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
disallow_any_generics = true
plugins = []

[[tool.mypy.overrides]]
module = "celery.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "celery"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.tasks.*"
disable_error_code = ["untyped-decorator"]

[[tool.mypy.overrides]]
module = "app.celery_app"
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

